requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
pybit==5.11.0
python-dotenv==1.1.1